extract canvas_widget, change some methods' names
